requirements.txt
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.19
Pillow==2.4.0
Werkzeug==0.9.4
argparse==1.2.1
gunicorn==18.0
itsdangerous==0.24
numpy==1.8.1
urwid==1.2.0
wsgiref==0.1.2
